🚑️ Fix IClickableMenu->drawHoverText patch and footsteps patch
